LayoutTests/accessibility/table-attributes.html

<html>
<head>
<script src="../resources/js-test.js"></script>
<script src="../resources/accessibility-helper.js"></script>
<script src="./resources/table-helper.js"></script>
</head>
<body>

<table id="testTable" align="center" border="1" cellpadding="5" cellspacing="0"
    summary="This is the summary text that should appear as a description">
    <caption> Example #1: Nested Stubs </caption>
    <tr>
        <th class="center" colspan="2" rowspan="2">Ruritanian<br> Population
            <br> Survey</th>
        <th class="center" rowspan="2">All<br> Genders</th>
        <th class="center" colspan="2">By Gender</th>
    </tr>
    <tr>
        <th class="center">Males</th>
        <th class="center">Females</th>
    </tr>
    <tr>
        <th align="left" rowspan="2">All Regions</th>
        <th>North</th>
        <td align="right">3333</td>
        <td align="right">1111</td>
        <td align="right">2222</td>
    </tr>
    <tr>
        <th>South</th>
        <td align="right">3333</td>
        <td align="right">1111</td>
        <td align="right">2222</td>
    </tr>
</table>

<script>
    if (window.accessibilityController) {
        let table = accessibilityController.accessibleElementById("testTable");
        axDebug(tableProperties(table));
    }
</script>
</body>
</html>
